migrate setTimer action
