+++ bigbluebutton-html5/imports/api/timer/server/methods.js
@@ -1,10 +1,8 @@
 import { Meteor } from 'meteor/meteor';
-import setTimer from './methods/setTimer';
 import getServerTime from './methods/getServerTime';
 import timerEnded from './methods/endTimer';
 
 Meteor.methods({
-  setTimer,
   getServerTime,
   timerEnded,
 });
