--- bigbluebutton-html5/imports/api/timer/server/methods/setTimer.js
@@ -1,25 +1,0 @@
-import { check } from 'meteor/check';
-import RedisPubSub from '/imports/startup/server/redis';
-import Logger from '/imports/startup/server/logger';
-import { extractCredentials } from '/imports/api/common/server/helpers';
-
-export default function setTimer(time) {
-  const REDIS_CONFIG = Meteor.settings.private.redis;
-  const CHANNEL = REDIS_CONFIG.channels.toAkkaApps;
-  const EVENT_NAME = 'SetTimerReqMsg';
-
-  try {
-    const { meetingId, requesterUserId } = extractCredentials(this.userId);
-    check(meetingId, String);
-    check(requesterUserId, String);
-    check(time, Number);
-
-    const payload = {
-      time,
-    };
-
-    RedisPubSub.publishUserMessage(CHANNEL, EVENT_NAME, meetingId, requesterUserId, payload);
-  } catch (err) {
-    Logger.error(`Setting timer: ${err}`);
-  }
-}
